fix(titanium-address-input): floating label bug on inputs in dialog if prefilled; refactor to use mwc-dialog
