+++ packages/titanium-address-input/package.json
@@ -20,11 +20,11 @@
   },
   "dependencies": {
     "@googlemaps/js-api-loader": "*",
-    "@leavittsoftware/titanium-dialog": "^3.6.5",
     "@material/mwc-button": "*",
     "@material/mwc-list": "*",
     "@material/mwc-select": "*",
     "@material/mwc-textfield": "*",
+    "@material/mwc-dialog": "*",
     "tslib": "^2.3.1"
   },
   "peerDependencies": {
